pull repl from command into its own unit
This allows adding some tests to tab completion (which is currently a
little broken).
