+++ lib/command.js
@@ -1,10 +1,10 @@
 // Generated by LiveScript 1.5.0
-var LiveScript, path, fs, util, prelude, each, breakList, lines, unlines, take, keys, filter, dasherize, map, ref$, parseOptions, generateHelp, nameFromPath, SourceNode, version, this$ = this, slice$ = [].slice;
+var LiveScript, path, fs, util, prelude, each, breakList, ref$, parseOptions, generateHelp, nameFromPath, SourceNode, version, this$ = this, slice$ = [].slice;
 LiveScript = require('..');
 path = require('path');
 fs = require('fs');
 util = require('util');
-prelude = require('prelude-ls'), each = prelude.each, breakList = prelude.breakList, lines = prelude.lines, unlines = prelude.unlines, take = prelude.take, keys = prelude.keys, filter = prelude.filter, dasherize = prelude.dasherize, map = prelude.map;
+prelude = require('prelude-ls'), each = prelude.each, breakList = prelude.breakList;
 ref$ = require('./options'), parseOptions = ref$.parse, generateHelp = ref$.generateHelp;
 nameFromPath = require('./util').nameFromPath;
 SourceNode = require('source-map').SourceNode;
@@ -12,7 +12,7 @@ version = LiveScript.VERSION;
 (function(it){
   return module.exports = it;
 })(function(args, arg$){
-  var ref$, say, sayWithTimestamp, warn, die, p, pp, ppp, fileExists, dasherizeVars, startsWith, o, positional, e, validMapValues, toInsert, that, filename, jsonCallback;
+  var ref$, say, sayWithTimestamp, warn, die, o, positional, e, validMapValues, toInsert, that, filename, jsonCallback;
   ref$ = arg$ != null
     ? arg$
     : {}, say = ref$.say, sayWithTimestamp = ref$.sayWithTimestamp, warn = ref$.warn, die = ref$.die;
@@ -23,40 +23,6 @@ version = LiveScript.VERSION;
     console.error(message);
     process.exit(1);
   });
-  p = function(){
-    var args, res$, i$, to$;
-    res$ = [];
-    for (i$ = 0, to$ = arguments.length; i$ < to$; ++i$) {
-      res$.push(arguments[i$]);
-    }
-    args = res$;
-    each(console.dir, args);
-  };
-  pp = function(x, showHidden, depth){
-    say(util.inspect(x, showHidden, depth, !process.env.NODE_DISABLE_COLORS));
-  };
-  ppp = function(it){
-    pp(it, true, null);
-  };
-  fileExists = function(path){
-    try {
-      fs.statSync(path);
-      return true;
-    } catch (e$) {}
-  };
-  dasherizeVars = function(str){
-    if (/^[a-z]/.exec(str)) {
-      return dasherize(str);
-    } else {
-      return str;
-    }
-  };
-  startsWith = function(str){
-    var this$ = this;
-    return function(it){
-      return it.indexOf(str) === 0;
-    };
-  };
   try {
     o = parseOptions(args);
     positional = o._;
@@ -128,7 +94,7 @@ version = LiveScript.VERSION;
       break;
     case !require('tty').isatty(0):
       say("LiveScript " + version + " - use 'lsc --help' for more information");
-      repl();
+      require('./repl')(o);
       break;
     default:
       compileStdin();
@@ -388,158 +354,6 @@ version = LiveScript.VERSION;
       say(l ? l.join(' ').replace(/\n/g, '\\n') : '');
     }
   }
-  function repl(){
-    var MAXHISTORYSIZE, historyFile, code, cont, rl, reset, _ttyWrite, prompt, that, vm, replCtx, server, ref$;
-    MAXHISTORYSIZE = 500;
-    historyFile = path.join(process.env.HOME, '/.lsc_history');
-    code = repl.infunc ? '  ' : '';
-    cont = 0;
-    rl = require('readline').createInterface(process.stdin, process.stdout);
-    reset = function(){
-      rl.line = code = '';
-      rl.prompt();
-      repl.inheredoc = false;
-    };
-    (_ttyWrite = rl._ttyWrite, rl)._ttyWrite = function(char){
-      if (char === '\n' || char === '>') {
-        cont += 1;
-      } else {
-        cont = 0;
-      }
-      return _ttyWrite.apply(this, arguments);
-    };
-    prompt = 'ls';
-    if (that = repeatString$('b', !!o.bare) + repeatString$('c', !!o.compile)) {
-      prompt += " -" + that;
-    }
-    try {
-      rl.history = lines(fs.readFileSync(historyFile, 'utf-8').trim());
-    } catch (e$) {}
-    if (LiveScript != null) {
-      LiveScript.history = rl.history;
-    }
-    if (!o.compile) {
-      module.paths = module.constructor._nodeModulePaths(module.filename = process.cwd() + '/repl');
-      vm = require('vm');
-      if (o.prelude) {
-        import$(global, prelude);
-      }
-      replCtx = {};
-      import$(replCtx, global);
-      replCtx.module = module;
-      replCtx.exports = exports;
-      replCtx.require = require;
-      replCtx.LiveScript = LiveScript;
-      replCtx.path = path;
-      replCtx.fs = fs;
-      replCtx.util = util;
-      replCtx.say = say;
-      replCtx.warn = warn;
-      replCtx.die = die;
-      replCtx.p = p;
-      replCtx.pp = pp;
-      replCtx.ppp = ppp;
-      server = (ref$ = clone$(require('repl').REPLServer.prototype), ref$.context = replCtx, ref$.commands = [], ref$.useGlobal = true, ref$.useColors = process.env.NODE_DISABLE_COLORS, ref$.eval = function(code, ctx, arg$, cb){
-        var res, e;
-        try {
-          res = vm.runInNewContext(code, ctx, 'repl');
-        } catch (e$) {
-          e = e$;
-        }
-        cb(e, res);
-      }, ref$);
-      rl.completer = function(line, cb){
-        var contextVars, matches;
-        contextVars = map(dasherizeVars, keys(server.context));
-        matches = filter(startsWith(line), contextVars);
-        return cb(null, [matches.length ? matches : contextVars, line]);
-      };
-    }
-    rl.on('SIGCONT', rl.prompt);
-    rl.on('SIGINT', function(){
-      if (this.line || code) {
-        say('');
-        reset();
-      } else {
-        this.close();
-      }
-    });
-    rl.on('close', function(){
-      say('');
-      return process.exit();
-    });
-    rl.on('line', function(it){
-      var isheredoc, ops, x, e;
-      if (it.match(/^$/)) {
-        repl.infunc = false;
-      }
-      if (it.match(/(\=|\~>|->|do|import|switch)\s*$/) || (it.match(/^!?(function|class|if|unless) /) && !it.match(/ then /))) {
-        repl.infunc = true;
-      }
-      if (((0 < cont && cont < 3) || repl.infunc) && !repl.inheredoc) {
-        code += it + '\n';
-        this.output.write(repeatString$('.', prompt.length) + '. ');
-        return;
-      } else {
-        isheredoc = it.match(/(\'\'\'|\"\"\")/g);
-        if (isheredoc && isheredoc.length % 2 === 1) {
-          repl.inheredoc = !repl.inheredoc;
-        }
-        if (repl.inheredoc) {
-          code += it + '\n';
-          rl.output.write(repeatString$('.', prompt.length) + '" ');
-          return;
-        }
-      }
-      repl.inheredoc = false;
-      if (!(code += it)) {
-        return reset();
-      }
-      try {
-        if (o.compile) {
-          say(LiveScript.compile(code, {
-            bare: o.bare
-          }));
-        } else {
-          ops = {
-            'eval': 'eval',
-            bare: true,
-            saveScope: LiveScript
-          };
-          if (code.match(/^\s*!?function/)) {
-            ops = {
-              bare: true
-            };
-          }
-          x = vm.runInNewContext(LiveScript.compile(code, ops), replCtx, 'repl');
-          if (x != null) {
-            replCtx._ = x;
-          }
-          pp(x);
-          if (typeof x === 'function') {
-            say(x);
-          }
-        }
-      } catch (e$) {
-        e = e$;
-        say(e);
-      }
-      reset();
-    });
-    process.on('uncaughtException', function(it){
-      say("\n" + ((it != null ? it.stack : void 8) || it));
-    });
-    process.on('exit', function(){
-      if (code && rl.output.isTTY) {
-        rl._ttyWrite('\r');
-      }
-      if (fileExists(historyFile)) {
-        fs.writeFileSync(historyFile, compose$(take(MAXHISTORYSIZE), unlines)(rl.history));
-      }
-    });
-    rl.setPrompt(prompt + "> ");
-    rl.prompt();
-  }
   function forkNode(){
     var ref$, args, lsArgs, ref1$, nodeArgs, this$ = this;
     ref$ = process.argv, args = slice$.call(ref$, 1);
@@ -561,23 +375,4 @@ function import$(obj, src){
   var own = {}.hasOwnProperty;
   for (var key in src) if (own.call(src, key)) obj[key] = src[key];
   return obj;
-}
-function repeatString$(str, n){
-  for (var r = ''; n > 0; (n >>= 1) && (str += str)) if (n & 1) r += str;
-  return r;
-}
-function clone$(it){
-  function fun(){} fun.prototype = it;
-  return new fun;
-}
-function compose$() {
-  var functions = arguments;
-  return function() {
-    var i, result;
-    result = functions[0].apply(this, arguments);
-    for (i = 1; i < functions.length; ++i) {
-      result = functions[i](result);
-    }
-    return result;
-  };
 }
